Improved frame drag & sliding/maximizing for 1.2.x
Most issues with dragging or clicking at the top of a non-native frame
should now be resolved, along with some sizing issues related to
maximizing.  Toggling maximize while a sidebar pane is active will
now apply to the most-recently-active pane in the main area, instead
of hiding it.

This was a major overhaul of the plugin's styling to support three
different frame styles (obsidian, hidden, and native) with and without
maximizing and sliding, and has only been tested on Windows.  It is
entirely possible I've missed a bug in a configuration or platform I
don't normally use (e.g. Mac, mobile, themes w/elaborate restyling,
etc.), so please open bug reports on Github if you experience any
issues.