Fix 1.7.5 not opening search w/hotkey + focus lock
Older versions of Obsidian focused the search box directly from
the search command, which allowed hotkey-based search to work
with Pane Relief's Focus Lock active.  Newer versions now do
the focusing in a way that is subject to Focus Lock, so this
release adds a workaround to emulate the old behavior.

Note: with the exception of taking focus away from popout windows,
Obsidian has fixed most of the problems in its UI that Focus Lock was
intended to prevent, so you may want to try turning off Focus Lock to
see if you still need it.  (The one other major use case for it is
keeping the "Automatically reveal active file" plugin from constantly
focusing the file explorer pane.)